+++ sources/core/Stride.Core.CompilerServices.Tests/TestHelper.cs
@@ -14,4 +14,10 @@ internal class TestHelper
         bool hasError = diagnostics.Any();
         Assert.False(hasError, $"The Test is valid and shouldn't throw Diagnostics. Thrown Diagnostics: {diagnostics.Select(x => x.Id)}");
     }
+    public static void ExpectDiagnosticsError(string sourceCode,string diagnosticID)
+    {
+        var diagnostics = CompilerUtils.CompileAndGetAnalyzerDiagnostics(sourceCode, CompilerUtils.AllAnalyzers);
+        bool hasError = diagnostics.Any();
+        Assert.False(hasError, $"The Test is invalid and should throw the '{diagnosticID}' Diagnostics. Thrown Diagnostics: {diagnostics.Select(x => x.Id)}");
+    }
 }
